% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_spatial_join.R
\name{sdf_spatial_join}
\alias{sdf_spatial_join}
\title{Spatial Join}
\usage{
sdf_spatial_join(left, right, latitude, longitude, polygon = "polygon",
  dropPolyAfterJoin = TRUE)
}
\arguments{
\item{left}{An object coercable to a Spark DataFrame.}

\item{right}{An object coercable to a Spark DataFrame.}

\item{latitude}{Latitude variable.}

\item{longitude}{Longitude variable.}

\item{polygon}{Variable representing the polygon to join.}

\item{dropPolyAfterJoin}{\code{logical} specifying whether
the polygon column (and the index if available) should be
dropped after the join. Defaults to \code{TRUE}.}
}
\value{
\code{tbl_spark} resulting from the spatial join.
}
\description{
Spatially joins two \code{tbl}.
}
\references{
\url{https://github.com/harsha2010/magellan}
\url{https://magellan.ghost.io/how-does-magellan-scale-geospatial-queries/}
}
