% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkgeo_register.R
\name{sparkgeo_register}
\alias{sparkgeo_register}
\title{Register sparkgeo and Spark User-Defined Functions (UDFs)}
\usage{
sparkgeo_register(sc)
}
\arguments{
\item{sc}{\code{\link[sparklyr]{spark_connection}} provided by sparklyr.}
}
\description{
Registers Magellan rules and custom UDFs with Spark
}
