% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkgeo_register_udfs.R
\name{sparkgeo_register_udfs}
\alias{sparkgeo_register_udfs}
\title{Register Spark User-Defined Functions (UDFs)}
\usage{
sparkgeo_register_udfs(sc)
}
\arguments{
\item{sc}{\code{\link[sparklyr]{spark_connection}} provided by sparklyr.}
}
\description{
Registers UDFs with Spark
}
